+++ libraries/Ethernet/IPAddress.cpp
@@ -42,9 +42,9 @@ bool IPAddress::operator==(const uint8_t* addr)
     return memcmp(addr, _address, sizeof(_address)) == 0;
 }
 
-long IPAddress::printTo(Print& p) const
+ssize_t IPAddress::printTo(Print& p) const
 {
-    long n = 0, t;
+    ssize_t n = 0, t;
     for (int i =0; i < 3; i++)
     {
         if ((t = p.print(_address[i], DEC)) > 0) n += t;
